fix(file system): divide mainScreen into parts for easier readability
