+++ app/src/main/kotlin/view/mainScreen/ThemeSwitcher.kt
@@ -1,4 +1,4 @@
-package view
+package view.mainScreen
 
 import androidx.compose.animation.core.AnimationSpec
 import androidx.compose.animation.core.animateDpAsState
@@ -23,6 +23,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import view.Theme
 
 
 @Composable
