+++ app/src/main/kotlin/view/mainScreen/resetView.kt
@@ -1,0 +1,28 @@
+package view.mainScreen
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Button
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.Text
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import viewmodel.MainScreenViewModel
+
+@Composable
+fun <V> resetGraphView(viewModel: MainScreenViewModel<V>) {
+    Button(
+        onClick = viewModel::resetGraphView,
+        enabled = true,
+        colors = ButtonDefaults.outlinedButtonColors(
+            backgroundColor = MaterialTheme.colorScheme.secondary
+        ),
+        modifier = Modifier.padding(horizontal = 4.dp)
+    ) {
+        Text(
+            text = "Reset default settings",
+            color = MaterialTheme.colorScheme.onSecondary
+        )
+    }
+}
