+++ app/src/main/kotlin/view/mainScreen/showLabels.kt
@@ -1,0 +1,58 @@
+package view.mainScreen
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Checkbox
+import androidx.compose.material.CheckboxDefaults
+import androidx.compose.material.Text
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import viewmodel.MainScreenViewModel
+
+@Composable
+fun <V> showVerticesLabels(viewModel: MainScreenViewModel<V>) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Checkbox(
+            checked = viewModel.showVerticesLabels.value,
+            onCheckedChange = { viewModel.showVerticesLabels.value = it },
+            colors = CheckboxDefaults.colors(
+                checkmarkColor = MaterialTheme.colorScheme.onSecondary,
+                checkedColor = MaterialTheme.colorScheme.secondary,
+            )
+        )
+        Text(
+            text = "Show vertices' labels",
+            fontSize = 16.sp,
+            color = MaterialTheme.colorScheme.onSurface,
+            modifier = Modifier.padding(horizontal = 4.dp)
+        )
+    }
+}
+
+@Composable
+fun <V> showEdgesLabels(viewModel: MainScreenViewModel<V>) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Checkbox(
+            checked = viewModel.showEdgesLabels.value,
+            onCheckedChange = { viewModel.showEdgesLabels.value = it },
+            colors = CheckboxDefaults.colors(
+                checkmarkColor = MaterialTheme.colorScheme.onSecondary,
+                checkedColor = MaterialTheme.colorScheme.secondary,
+            )
+        )
+        Text(
+            text = "Show edges' weights",
+            fontSize = 16.sp,
+            color = MaterialTheme.colorScheme.onSurface,
+            modifier = Modifier.padding(horizontal = 4.dp)
+        )
+    }
+}
